+++ docs-demo/advanced/virtual/VirtualY.vue
@@ -20,7 +20,12 @@ const dataSource = [
     { name: `Tom`, age: 20, address: `Shanghai`, gender: 'male' },
     { name: `Lucy`, age: 22, address: `Guangzhou`, gender: 'female' },
     { name: `Lily`, age: 24, address: `Shenzhen`, gender: 'female' },
-    ...new Array(5000).fill(0).map((_, i) => ({ name: `Jack${i}`, age: 18, address: `Beijing Forbidden City `, gender: 'male' })),
+    ...new Array(5000).fill(0).map((_, i) => ({
+        name: `Jack${i}`,
+        age: 18,
+        address: `Beijing Forbidden City `,
+        gender: 'male',
+    })),
 ];
 </script>
 <template>