+++ docs-demo/basic/fixed/FixedVirtual.vue
@@ -16,11 +16,46 @@ const columns: StkTableColumn<any>[] = [
     { title: 'Website', dataIndex: 'website', width: 100 },
 ];
 const dataSource = ref([
-    { key: '1', name: 'John Brown', age: 32, address: 'New York No. 1 Lake Park', gender: 'male', email: 'john@example.com' },
-    { key: '2', name: 'Jim Green', age: 42, address: 'London No. 1 Lake Park', gender: 'male', email: 'jim@example.com' },
-    { key: '3', name: 'Joe Black', age: 32, address: 'Sidney No. 1 Lake Park', gender: 'male', email: 'joe@example.com' },
-    { key: '4', name: 'Jim Red', age: 32, address: 'London No. 2 Lake Park', gender: 'male', email: 'jim@example.com' },
-    { key: '5', name: 'Jake White', age: 32, address: 'New York No. 2 Lake Park', gender: 'male', email: 'jake@example.com' },
+    {
+        key: '1',
+        name: 'John Brown',
+        age: 32,
+        address: 'New York No. 1 Lake Park',
+        gender: 'male',
+        email: 'john@example.com',
+    },
+    {
+        key: '2',
+        name: 'Jim Green',
+        age: 42,
+        address: 'London No. 1 Lake Park',
+        gender: 'male',
+        email: 'jim@example.com',
+    },
+    {
+        key: '3',
+        name: 'Joe Black',
+        age: 32,
+        address: 'Sidney No. 1 Lake Park',
+        gender: 'male',
+        email: 'joe@example.com',
+    },
+    {
+        key: '4',
+        name: 'Jim Red',
+        age: 32,
+        address: 'London No. 2 Lake Park',
+        gender: 'male',
+        email: 'jim@example.com',
+    },
+    {
+        key: '5',
+        name: 'Jake White',
+        age: 32,
+        address: 'New York No. 2 Lake Park',
+        gender: 'male',
+        email: 'jake@example.com',
+    },
     // 再加100条数据
     ...new Array(100).fill(0).map((_, i) => ({
         key: `${i + 6}`,
@@ -34,5 +69,13 @@ const dataSource = ref([
 </script>
 
 <template>
-    <StkTable style="height: 200px" row-key="key" virtual virtual-x fixed-col-shadow :columns="columns" :data-source="dataSource"></StkTable>
+    <StkTable
+        style="height: 200px"
+        row-key="key"
+        virtual
+        virtual-x
+        fixed-col-shadow
+        :columns="columns"
+        :data-source="dataSource"
+    ></StkTable>
 </template>