+++ docs-demo/basic/row-height/RowHeight.vue
@@ -23,10 +23,22 @@ const dataSource = ref([
     { name: `Tom`, age: 20, address: `Shanghai`, gender: 'male' },
     { name: `Lucy`, age: 22, address: `Guangzhou`, gender: 'female' },
     { name: `Lily`, age: 24, address: `Shenzhen`, gender: 'female' },
-    ...new Array(50).fill(0).map((_, i) => ({ name: `Jack${i}`, age: 18, address: `Beijing Forbidden City `, gender: 'male' })),
+    ...new Array(50).fill(0).map((_, i) => ({
+        name: `Jack${i}`,
+        age: 18,
+        address: `Beijing Forbidden City `,
+        gender: 'male',
+    })),
 ]);
 </script>
 <template>
     <CheckItem v-model="virtual" text="开启虚拟列表virtual"></CheckItem>
-    <StkTable style="height: 200px" :virtual="virtual" row-height="40" header-row-height="50" :columns="columns" :data-source="dataSource"></StkTable>
+    <StkTable
+        style="height: 200px"
+        :virtual="virtual"
+        row-height="40"
+        header-row-height="50"
+        :columns="columns"
+        :data-source="dataSource"
+    ></StkTable>
 </template>