+++ docs-src/.vitepress/config.ts
@@ -91,7 +91,7 @@ export default defineConfig({
                   { text: '序号列', link: '/table/basic/seq', },
                   { text: '空数据', link: '/table/basic/empty' },
                   { text: '行、列唯一键', link: '/table/basic/key', },
-                  { text: '滚动条样式', link: '/table/basic/scrollbar-style', },
+                  { text: '滚动条', link: '/table/basic/scrollbar', },
                   { text: 'table-layout:fix', link: '/table/basic/fixed-mode', },
                   { text: '按行滚动', link: '/table/basic/scroll-row-by-row', },
                 ]