+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.11
+* Feature
+  - add: Support `Home` `End` key to scroll table.
+* Bugfix
+  - fix: theme `light`, not append class `light` to stk-table root element.
+
 ## 0.6.10
 * Optimize 
   - Optimize scroll performance.(Fix: Execute code branches that do not meet expectations, when scrolling y)